requirements.txt
anthropic>=0.54.0
httpx>=0.27.0
apify-client>=1.8.0
python-dotenv>=1.0.0
